Columns splitted in several (5) tables
