+++ tasks/ca/statcan/census2016/data.py
@@ -1,8 +1,14 @@
 from luigi import Task, Parameter, LocalTarget, WrapperTask
 from tasks.ca.statcan.geo import (GEOGRAPHIES, GeographyColumns, Geography,
                                   GEO_CT, GEO_PR, GEO_CD, GEO_CSD, GEO_CMA, GEO_DA, GEO_FSA)
-from tasks.base_tasks import DownloadUnzipTask, RepoFile
+from tasks.ca.statcan.census2016.cols_census import CensusColumns
+
+from tasks.base_tasks import DownloadUnzipTask, RepoFile, TempTableTask
 from tasks.util import copyfile
+from tasks.meta import current_session
+from lib.logger import get_logger
+
+LOGGER = get_logger(__name__)
 
 GEOGRAPHY_CODES = {
     GEO_PR: '044',
@@ -32,8 +38,30 @@ class DownloadData(DownloadUnzipTask):
         copyfile(self.input().path, '{output}.zip'.format(output=self.output().path))
 
 
-class ImportData(WrapperTask):
+class ImportData(TempTableTask):
     resolution = Parameter()
+    table = Parameter()
 
     def requires(self):
-        return DownloadData(geocode=GEOGRAPHY_CODES[self.resolution])
+        return {
+            'data': DownloadData(geocode=GEOGRAPHY_CODES[self.resolution]),
+            'columns': CensusColumns(table=self.table),
+        }
+
+    def run(self):
+        _input = self.input()
+
+        columns = ['{} NUMERIC'.format(col) for col in _input['columns'].keys()]
+        session = current_session()
+        query = '''
+                CREATE TABLE {output} (
+                    geom_id TEXT,
+                    {columns}
+                )
+                '''.format(
+                    output=self.output().table,
+                    columns=','.join(columns),
+                )
+
+        LOGGER.debug(query)
+        session.execute(query)
